style(mkpasswd): fix audit issues
- Move version before sha256
- Use revision instead of underscore in version
- Use modifier unless with &&
- Replace on_macos with if OS.mac?
- Add final newline

+++ Formula/mkpasswd.rb
@@ -2,8 +2,9 @@ class Mkpasswd < Formula
   desc "Generate hashed passwords (mkpasswd from whois package)"
   homepage "https://github.com/rfc1036/whois"
   url "https://github.com/rfc1036/whois/archive/refs/tags/v5.6.6.tar.gz"
+  version "5.6.6"
+  revision 14
   sha256 "43d3b3cc64c75e8bd10aee6feff3906e9488ed335076d206e70f3b25bf644969"
-  version "5.6.6_14"
   license "GPL-2.0-or-later"
 
   depends_on "openssl@3"
@@ -16,9 +17,7 @@ class Mkpasswd < Formula
     inreplace "utils.c", /\A/, "#include <string.h>\n\n"
 
     # Generate version.h if missing (required for mkpasswd.c)
-    unless File.exist?("version.h")
-      File.write("version.h", "#define VERSION \"#{version}\"\n")
-    end
+    File.write("version.h", "#define VERSION \"#{version}\"\n") unless File.exist?("version.h")
 
     # Patch Makefile for pkg-config libcrypto
     inreplace "Makefile", /^else$/, <<~EOS
@@ -27,7 +26,7 @@ class Mkpasswd < Formula
       else
     EOS
 
-    on_macos do
+    if OS.mac?
       ENV.append "LDFLAGS", "-L/usr/lib -liconv"
     end
